V0.0.0 Eliot's First Commit
Eliot is still under heavy development, and is subject to bugs, crashes, or errors. (It should be fine for most users!)

Code documentation and commenting will be added in a later commit.

 This is my first time working with LLMs and seeing the process of which prompt engineering, chunking, behavioral temperatures, and tokenization interact to provide different results (output).

V0.0.0 includes the core of Eliot and it's overall work flow in a easy-to-use local GUI (website).

Stay tuned for updates and refinement!